Moves Pub/Sub Spring Integration to pubsub folder
In the future, other Spring Integrations can be added to other folders
(e.g., Storage to storage folder).
